mobok_back/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.12.5
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.7
django-cors-headers==3.7.0
django-filter==2.4.0
django-webpack-loader==0.7.0
djangorestframework==3.12.2
filelock==3.0.12
gunicorn==20.0.4
pipenv==2020.11.15
psycopg2-binary==2.8.6
pytz==2021.1
selenium==3.141.0
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.4
virtualenv==20.4.2
virtualenv-clone==0.5.4
whitenoise==5.2.0
